CommunityConfiguration: Remove incorrectly declared i18n messages
This removes the following two messages:
* communityconfiguration-editor-form-submit-button-text
* communityconfiguration-summary

Both removed messages are in "i18n namespace" that should
be declared from CommunityConfiguration, not other extensions.
Only messages starting with "communityconfiguration-PROVIDER-"
should be defined from other extensions. In AutoModerator's
case, this means "communityconfiguration-automoderator-*"
for now.

The first message is already declared in CommunityConfiguration
(and used), with the same content. This results in a conflict
at the TranslateWiki level, which is how this issue was noticed.

The second message is not declared in CommunityConfiguration,
but it is also not used anywhere as far as I can tell,
so removing it should be fine. If it is needed in the future,
it would belong to CommunityConfiguration itself, rather than here.

Bug: T374084
Change-Id: I2992ce1a45be93aaea48388f6285d969d575aa6d