Remove AutoModeratorConfigLoaderStaticTrait
This is not a real trait. Both its implemented methods are static,
and are shortcuts to accessing services from the container.
Traits are intended to share non-static methods between classes.

Note that the static methods were actually called as non-static.
While PHP allows it, it's confusing. Replace the calls with calls
to the service container.

Change-Id: Iaae29d4550397b947843acde157ed0941cbf9948