Add CommunityConfiguration defaults file and Schema test
This defaults file is intended to be used with CommunityConfiguration's
"EmergencyShutdown mode". In the unlikely case that there is a
fundamental problem with CommunityConfiguration in production (e.g. a
validation error causing a site outage, making it impossible to change
the configuration), then SREs can set a config-option to short-circuit
CommunityConfiguration. This disables as much of its code as possible,
especially loading config from a wikipage and parsing its defaults.
Instead, these pre-generated defaults will be used.

The test asserts that the defaults file is in sync with what is defined
in the Schema. In the future, it is planned to added additional checks
to the SchemaProviderTestCase to make sure that all Schemas are
sensible. For example that min < max, that top-level properties _have_
defaults, etc.

Bug: T367516
Bug: T374428
Depends-On: I2e2ac0b3fc18ee0296b49e2a225e8f631ce84a12
Change-Id: Ic99b6ab0e9f7717fa286822aba76b9afe456f1a1