Improve handling of reverts
$revertTags included "mw-reverted". However, this tag does not
indicate that the edit is a reverted, but that it was reverted
(perhaps someone could overtake AutoMod).

There is no point in dealing with such edits (why revert an edit
we know was reverted), so make it another pre-check step. Also,
simplify the code with array_intersect to avoid a loop. This
can avoid the parent revision lookup for most edit checks.

Change-Id: I131cbc36abe03089832a2a57ce39e47b2233ceff