Change AutoModerator's edit summary
Before, AutoModerator used to undo revisions and we had edit summaries
for logged-in and logged-out users. We have since changed AutoModerator's
behavior to rollback revisions, so we needed to change the edit
summary.

Since the edit summary will no longer be used, the undoSummary
parameter in the RevisionCheck class was deleted.

Bug: T385875
Change-Id: I63946403db59c004aa962262390b3105b4c099c0